requirements.txt
# Web3相关依赖
web3==6.15.1
eth-account==0.11.0
eth-utils==2.3.1

# 开发工具
python-dotenv==1.0.1
pytest==8.1.1
black==24.2.0
flake8==7.0.0
pre-commit==3.6.2
